+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/windmill/state/WindmillCombiningState.java
@@ -27,11 +27,13 @@ import org.apache.beam.runners.core.StateNamespace;
 import org.apache.beam.runners.core.StateTag;
 import org.apache.beam.runners.core.StateTags;
 import org.apache.beam.runners.dataflow.worker.windmill.Windmill;
+import org.apache.beam.runners.dataflow.worker.windmill.state.WindmillStateCache.ForKeyAndFamily;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.state.BagState;
 import org.apache.beam.sdk.state.CombiningState;
 import org.apache.beam.sdk.state.ReadableState;
 import org.apache.beam.sdk.transforms.Combine;
+import org.apache.beam.sdk.transforms.Combine.CombineFn;
 import org.apache.beam.vendor.grpc.v1p69p0.com.google.protobuf.ByteString;
 import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.base.Supplier;
 import org.apache.beam.vendor.guava.v32_1_2_jre.com.google.common.collect.Iterables;
@@ -55,11 +57,12 @@ class WindmillCombiningState<InputT, AccumT, OutputT> extends WindmillState
       StateTag<CombiningState<InputT, AccumT, OutputT>> address,
       String stateFamily,
       Coder<AccumT> accumCoder,
-      Combine.CombineFn<InputT, AccumT, OutputT> combineFn,
-      WindmillStateCache.ForKeyAndFamily cache,
-      boolean isNewKey) {
+      CombineFn<InputT, AccumT, OutputT> combineFn,
+      ForKeyAndFamily cache,
+      boolean isNewKey,
+      WindmillStateTagUtil windmillStateTagUtil) {
     StateTag<BagState<AccumT>> internalBagAddress = StateTags.convertToBagTagInternal(address);
-    ByteString encodeKey = WindmillStateUtil.encodeKey(namespace, internalBagAddress);
+    ByteString encodeKey = windmillStateTagUtil.encodeKey(namespace, internalBagAddress);
 
     this.bag =
         cache