+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigquery/providers/BigQueryStorageWriteApiSchemaTransformProvider.java
@@ -315,6 +315,9 @@ public class BigQueryStorageWriteApiSchemaTransformProvider
       if (!Strings.isNullOrEmpty(configuration.getKmsKey())) {
         write = write.withKmsKey(configuration.getKmsKey());
       }
+      if (configuration.getBigLakeConfiguration() != null) {
+        write = write.withBigLakeConfiguration(configuration.getBigLakeConfiguration());
+      }
       if (this.testBigQueryServices != null) {
         write = write.withTestServices(testBigQueryServices);
       }